requirements.txt
feedparser
vaderSentiment
python-dotenv
